LPS-105380 Rename class to match implemented class ModelIndexerWriterContributor
