+++ modules/apps/users-admin/users-admin-impl/src/main/java/com/liferay/users/admin/internal/search/spi/model/index/contributor/UserModelIndexerWriterContributor.java
@@ -12,28 +12,29 @@
  * details.
  */
 
-package com.liferay.account.internal.search.spi.model.index.contributor;
+package com.liferay.users.admin.internal.search.spi.model.index.contributor;
 
-import com.liferay.account.model.AccountEntry;
-import com.liferay.account.service.AccountEntryLocalService;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.search.batch.BatchIndexingActionable;
 import com.liferay.portal.search.batch.DynamicQueryBatchIndexingActionableFactory;
 import com.liferay.portal.search.spi.model.index.contributor.ModelIndexerWriterContributor;
 import com.liferay.portal.search.spi.model.index.contributor.helper.ModelIndexerWriterDocumentHelper;
+import com.liferay.users.admin.internal.search.ContactBatchReindexer;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Drew Brokke
+ * @author Luan Maoski
  */
 @Component(
 	immediate = true,
-	property = "indexer.class.name=com.liferay.account.model.AccountEntry",
+	property = "indexer.class.name=com.liferay.portal.kernel.model.User",
 	service = ModelIndexerWriterContributor.class
 )
-public class AccountEntryModelIndexWriterContributor
-	implements ModelIndexerWriterContributor<AccountEntry> {
+public class UserModelIndexerWriterContributor
+	implements ModelIndexerWriterContributor<User> {
 
 	@Override
 	public void customize(
@@ -41,27 +42,39 @@ public class AccountEntryModelIndexWriterContributor
 		ModelIndexerWriterDocumentHelper modelIndexerWriterDocumentHelper) {
 
 		batchIndexingActionable.setPerformActionMethod(
-			(AccountEntry accountEntry) -> batchIndexingActionable.addDocuments(
-				modelIndexerWriterDocumentHelper.getDocument(accountEntry)));
+			(User user) -> {
+				if (!user.isDefaultUser()) {
+					batchIndexingActionable.addDocuments(
+						modelIndexerWriterDocumentHelper.getDocument(user));
+				}
+			});
 	}
 
 	@Override
 	public BatchIndexingActionable getBatchIndexingActionable() {
 		return dynamicQueryBatchIndexingActionableFactory.
 			getBatchIndexingActionable(
-				_accountEntryLocalService.getIndexableActionableDynamicQuery());
+				userLocalService.getIndexableActionableDynamicQuery());
 	}
 
 	@Override
-	public long getCompanyId(AccountEntry accountEntry) {
-		return accountEntry.getCompanyId();
+	public long getCompanyId(User user) {
+		return user.getCompanyId();
 	}
+
+	@Override
+	public void modelIndexed(User user) {
+		contactBatchReindexer.reindex(user.getUserId(), user.getCompanyId());
+	}
+
+	@Reference
+	protected ContactBatchReindexer contactBatchReindexer;
 
 	@Reference
 	protected DynamicQueryBatchIndexingActionableFactory
 		dynamicQueryBatchIndexingActionableFactory;
 
 	@Reference
-	private AccountEntryLocalService _accountEntryLocalService;
+	protected UserLocalService userLocalService;
 
 }
